+++ R/Utilities.R
@@ -639,6 +639,19 @@ createOrderKey <- function(x, split = "/") {
 	if(is.na(firstNonNA)) {
 		firstNonNA <- x[min(which(!is.na(x)))]
 	}
+	# If the first element is coercable to numierc, try converting the entire vector to numeric, and check that no NAs were generated:
+	if(!is.na(suppressWarnings(as.numeric(x[1])))) {
+		numberOfNAs <- sum(is.na(x))
+		xnumeric <- suppressWarnings(as.numeric(x))
+		if(sum(is.na(xnumeric)) > numberOfNAs) {
+			return(x)
+		}
+		else {
+			return(xnumeric)
+		}
+	}
+	
+	# Split by the 'split' argument:
 	if(!grepl(split, firstNonNA)) {
 		return(x)
 	}
@@ -658,7 +671,7 @@ createOrderKey <- function(x, split = "/") {
 	# Only accept if all elements can be converted to numeric:
 	#if(any(is.na(splittedDT))) {
 	
-	# Acccpet if any oof the values are not NA:
+	# Acccpet if any of the values are not NA:
 	if(all(is.na(splittedDT))) {
 		return(x)
 	}
@@ -666,8 +679,9 @@ createOrderKey <- function(x, split = "/") {
 	# Convert to integer ranks:
 	#splittedDT[, names(splittedDT) := lapply(.SD, function(y) match(y, sort(unique(y))))]
 	# Replicate data.table's soring which happend in C-locale (see ?data.table::setorderv):
-	splittedDT[, names(splittedDT) := lapply(.SD, function(y) match(y, stringi::stri_sort(unique(y), locale = "C")))]
-	
+	#splittedDT[, names(splittedDT) := lapply(.SD, function(y) match(y, stringi::stri_sort(unique(y), locale = "C")))]
+	#splittedDT[, names(splittedDT) := lapply(.SD, function(y) match(y, stringr::str_sort(unique(y), locale = "C")))]
+	splittedDT[, names(splittedDT) := lapply(.SD, function(y) match(y, stringi::stri_sort(unique(y), locale = "en_US_POSIX")))]
 
 	# Count the maximum number of digits for each column, and multiply by the cummulative number of digits:
 	numberOfDigits <- splittedDT[, lapply(.SD, max)]