+++ tests/testthat/test-readXmlFile.R
@@ -76,3 +76,15 @@ for(item in icesFiles) {
 	# There should be minimal differences (in the Survey table only)
 	expect_true(length(all.equal(icesDataA, icesDataB)) <= 1)
 }
+
+# Zipped files:
+context("test-readXmlFile: Zipped acoustic file")
+example <- system.file("testresources", "echosounder_2020821.zip", package="RstoxData")
+parsedAcousticZip <- readXmlFile(example)
+expect_equal(parsedAcousticZip$sa$sa[1], 67.7185200)
+
+context("test-readXmlFile: Zipped biotic file")
+example <- system.file("testresources", "biotic_2020821.zip", package="RstoxData")
+parsedBioticZip <- readXmlFile(example)
+expect_equal(parsedBioticZip$individual$individualweight[1], 0.022)
+