+++ tests/testthat/test-sort.R
@@ -1,0 +1,9 @@
+context("test-sort")
+
+exampleFile <- system.file("testresources","biotic_2020821.zip", package="RstoxData")
+exampleData <- StoxBiotic(ReadBiotic(exampleFile))
+
+expecteFile <- system.file("testresources","biotic_2020821.rds", package="RstoxData")
+expectedData <- readRDS(expecteFile)
+
+expect_equal(exampleData, expectedData)